requirements.txt
google.generativeai
steamlit
opencv-python
python-docx
cv2
pillow
